face_rating/requirements.txt
Django==4.2.6
cmake==3.25.2
dlib==19.24.2
face-recognition==1.3.0
imutils==0.5.4
opencv-python==4.10.0.84
numpy==1.26.4
pillow==10.4.0